[수정] API baseURL 변경 및 응답 모델 변경 - baseURL 변경. - 응답 모델 데이터 타입 변경

+++ MobileTracker/MobileTracker/Model/Response.swift
@@ -26,7 +26,7 @@ struct BlockStruct: Codable {
         var version: String
         var prevBlockHash: String
         var merkleTreeRootHash: String
-        var timeStamp: UInt64
+        var timeStamp: Double
         var confirmedTransactionList: [TransactionInfo]
         var blockHash: String
         var height: Int
